Replace prop drilling with some of the redux store states so far
